chore: renamed directories
