--- document-parent/mongodb-parent/deployment/src/main/java/io/quarkiverse/jnosql/document/mongodb/deployment/Processor.java
@@ -1,24 +1,0 @@
-package io.quarkiverse.jnosql.document.mongodb.deployment;
-
-import org.eclipse.jnosql.databases.mongodb.communication.QuarkusMongoDBDocumentConfiguration;
-
-import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
-import io.quarkus.deployment.annotations.BuildProducer;
-import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.deployment.builditem.FeatureBuildItem;
-
-class Processor {
-
-    private static final String FEATURE = "jnosql-document-mongodb";
-
-    @BuildStep
-    FeatureBuildItem feature() {
-        return new FeatureBuildItem(FEATURE);
-    }
-
-    @BuildStep
-    void build(BuildProducer<AdditionalBeanBuildItem> additionalBeanProducer) {
-        additionalBeanProducer.produce(AdditionalBeanBuildItem.unremovableOf(QuarkusMongoDBDocumentConfiguration.class));
-    }
-
-}
