--- document-parent/mongodb-parent/integration-tests/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-jnosql.document.database=test
