--- document-parent/mongodb-parent/integration-tests/src/test/java/io/quarkiverse/jnosql/document/mongodb/it/JNoSQLResourceIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.jnosql.document.mongodb.it;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class JNoSQLResourceIT extends JNoSQLResourceTest {
-}
