--- document-parent/mongodb-parent/pom.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-    <parent>
-        <groupId>io.quarkiverse.jnosql</groupId>
-        <artifactId>quarkus-jnosql-document-parent</artifactId>
-        <version>3.3.5-SNAPSHOT</version>
-    </parent>
-    
-    <packaging>pom</packaging>
-    <artifactId>quarkus-jnosql-document-mongodb-parent</artifactId>
-    <name>Quarkus JNoSQL - Document - MongoDB - Parent</name>
-
-    <modules>
-        <module>runtime</module>
-        <module>deployment</module>
-    </modules>
-
-    <profiles>
-        <profile>
-            <id>it</id>
-            <activation>
-                <property>
-                    <name>performRelease</name>
-                    <value>!true</value>
-                </property>
-            </activation>
-            <modules>
-                <module>integration-tests</module>
-            </modules>
-        </profile>
-    </profiles>
-
-</project>
