--- document-parent/mongodb-parent/runtime/pom.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.quarkiverse.jnosql</groupId>
-        <artifactId>quarkus-jnosql-document-mongodb-parent</artifactId>
-        <version>3.3.5-SNAPSHOT</version>
-    </parent>
-    <artifactId>quarkus-jnosql-document-mongodb</artifactId>
-    <name>Quarkus JNoSQL - Document - MongoDB - Runtime</name>
-    <description>Quarkus JNoSQL Document MongoDB Runtime provides runtime support for integrating MongoDB
-        document-oriented databases with Quarkus applications. This extension facilitates seamless integration, enabling
-        efficient storage, retrieval, and management of document-oriented data within Quarkus projects, enhancing
-        developer productivity.
-    </description>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.quarkiverse.jnosql</groupId>
-            <artifactId>quarkus-jnosql-document</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-mongodb-client</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.jnosql.databases</groupId>
-            <artifactId>jnosql-mongodb</artifactId>
-        </dependency>
-    </dependencies>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-extension-maven-plugin</artifactId>
-                <version>${quarkus.version}</version>
-                <executions>
-                    <execution>
-                        <phase>compile</phase>
-                        <goals>
-                            <goal>extension-descriptor</goal>
-                        </goals>
-                        <configuration>
-                            <deployment>${project.groupId}:${project.artifactId}-deployment:${project.version}
-                            </deployment>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <annotationProcessorPaths>
-                        <path>
-                            <groupId>io.quarkus</groupId>
-                            <artifactId>quarkus-extension-processor</artifactId>
-                            <version>${quarkus.version}</version>
-                        </path>
-                    </annotationProcessorPaths>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
