--- document-parent/mongodb-parent/runtime/src/main/java/org/eclipse/jnosql/databases/mongodb/communication/QuarkusMongoDBDocumentConfiguration.java
@@ -1,22 +1,0 @@
-package org.eclipse.jnosql.databases.mongodb.communication;
-
-import jakarta.inject.Inject;
-import jakarta.inject.Singleton;
-
-import org.eclipse.jnosql.communication.Settings;
-import org.eclipse.jnosql.communication.semistructured.DatabaseConfiguration;
-
-import com.mongodb.client.MongoClient;
-
-@Singleton
-public class QuarkusMongoDBDocumentConfiguration implements DatabaseConfiguration {
-
-    @Inject
-    protected MongoClient client;
-
-    @Override
-    public MongoDBDocumentManagerFactory apply(Settings settings) throws NullPointerException {
-        return new MongoDBDocumentManagerFactory(this.client);
-    }
-
-}
