--- document-parent/mongodb-parent/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,9 +1,0 @@
-name: JNoSQL Document MongoDB
-#description: Do something useful.
-metadata:
-#  keywords:
-#    - jnosql
-#  guide: https://quarkiverse.github.io/quarkiverse-docs/jnosql/dev/
-#  categories:
-#    - "miscellaneous"
-#  status: "preview"
