+++ mongodb/integration-tests/src/main/java/io/quarkiverse/jnosql/mongodb/it/People.java
@@ -1,0 +1,9 @@
+package io.quarkiverse.jnosql.mongodb.it;
+
+import jakarta.data.repository.Repository;
+
+import org.eclipse.jnosql.mapping.NoSQLRepository;
+
+@Repository
+public interface People extends NoSQLRepository<Person, String> {
+}
