+++ mongodb/integration-tests/src/main/java/io/quarkiverse/jnosql/mongodb/it/PersonRecord.java
@@ -1,0 +1,25 @@
+package io.quarkiverse.jnosql.mongodb.it;
+
+import java.util.List;
+import java.util.UUID;
+
+import jakarta.nosql.Column;
+import jakarta.nosql.Entity;
+import jakarta.nosql.Id;
+
+@Entity
+public record PersonRecord(@Id String id,
+        @Column String name,
+        @Column List<String> phones) {
+
+    public static PersonRecord randomPerson() {
+        var person = new PersonRecord(
+                UUID.randomUUID().toString(),
+                UUID.randomUUID().toString(),
+                List.of(
+                        UUID.randomUUID().toString(),
+                        UUID.randomUUID().toString(),
+                        UUID.randomUUID().toString()));
+        return person;
+    }
+}
