+++ mongodb/integration-tests/src/main/resources/META-INF/beans.xml
