+++ mongodb/integration-tests/src/test/java/io/quarkiverse/jnosql/document/mongodb/it/JNoSQLResourceIT.java
