+++ mongodb/runtime/src/main/java/org/eclipse/jnosql/databases/mongodb/communication/QuarkusMongoDBDocumentConfiguration.java
