--- quarkus-jnosql-mongodb-parent/integration-tests/src/main/java/io/quarkiverse/jnosql/document/mongodb/it/PeopleRecord.java
@@ -1,9 +1,0 @@
-package io.quarkiverse.jnosql.document.mongodb.it;
-
-import jakarta.data.repository.Repository;
-
-import org.eclipse.jnosql.mapping.NoSQLRepository;
-
-@Repository
-public interface PeopleRecord extends NoSQLRepository<PersonRecord, String> {
-}
